+++ CHANGELOG.md
@@ -25,6 +25,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * Added unit tests for all file parsing functions in `utils.files`.
   * Reduced code duplication in the `utils.files.parse_fixed_width_filenames`
     and `utils.files.parse_delimited_filenames` functions
+  * Added ability to set Meta data using `meta['data_var', 'label'] = value`
+    structure
 * Bug Fix
   * Fixed `utils.files.parse_fixed_width_filenames` output for empty file list
 * Maintenance